migrate frontend to new version
