--- picasso/templates/PartialOcclusion.html
@@ -1,40 +1,0 @@
-<!--
-Copyright (c) 2017 Merantix GmbH
-All rights reserved. This program and the accompanying materials
-are made available under the terms of the Eclipse Public License v1.0
-which accompanies this distribution, and is available at
-http://www.eclipse.org/legal/epl-v10.html
-
-Contributors:
-   Ryan Henderson - initial API and implementation and/or initial
-   documentation
--->
-{% extends "result.html" %}
-{% block vis %}
-	<table class="table table-sm table-striped">
-		<tbody>
-			{% for result in results %}
-				<tr>
-					<td align="center"><b>{{ result.filename }}</b></td>
-					<td align="center"><b>Occlusion Grid</b></td>
-					{% for prob in result.predict_probs %}
-						<td align="center"><b>{{ prob.name }}: {{ prob.prob }}</b></td>
-					{% endfor %}
-				</tr>
-				<tr>
-					<td align="center">
-						<img src="inputs/{{ result.filename }}" style="width:244px;height:244px;"/>
-					</td>
-					<td align="center">
-						<img src="outputs/{{ result.processed_input_file_name }}" style="width:244px;height:244px;"/>
-					</td>
-					{% for filename in result.output_file_names %}
-						<td align="center">
-							<img src="outputs/{{ filename }}" style="width:244px;height:244px;"/>
-						</td>
-					{% endfor %}
-				</tr>
-			{% endfor %}
-		</tbody>
-  </table>
-{% endblock %}
