--- picasso/templates/layout.html
@@ -1,43 +1,0 @@
-<!--
-Copyright (c) 2017 Merantix GmbH
-All rights reserved. This program and the accompanying materials
-are made available under the terms of the Eclipse Public License v1.0
-which accompanies this distribution, and is available at
-http://www.eclipse.org/legal/epl-v10.html
-
-Contributors:
-   Ryan Henderson - initial API and implementation and/or initial
-   documentation
-   Josh Chen - refactor and class config
--->
-<!doctype html>
-<head>
-  <title>{{ app_state.app_title }}</title>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
-  <link rel=stylesheet type=text/css href="{{ url_for('static',
-  filename='style.css') }}">
-	<link rel="shortcut icon" href="{{ url_for('static', filename='favicon.ico') }}">
-  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
-  <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
-</head>
-<div class="container-fluid">
-	<h1><a href="https://github.com/merantix/picasso">{{ app_state.app_title }}</a>
-		<small class="text-muted"> by
-		 	<a href="http://www.merantix.com">Merantix</a>
-		</small>
-	</h1>
-  <div class=info>
-		<h4><b>Current backend:</b> {{ app_state.model_name }}</h4>
-    {% if app_state.latest_ckpt_name is defined %}
-		<h4><b>Current checkpoint:</b> {{ app_state.latest_ckpt_name }}</h4>
-    {% endif %}
-    {% if app_state.latest_ckpt_time is defined %}
-		<h4><b>Last updated:</b> {{ app_state.latest_ckpt_time }}</h4>
-    {% endif %}
-    <a href="{{ url_for('picasso.landing') }}" class="btn btn-primary" role="button">Start over</a>
-  </div>
-  <hr>
-  {% block body %}{% endblock %}
-</div>
