--- picasso/templates/result.html
@@ -1,30 +1,0 @@
-<!--
-Copyright (c) 2017 Merantix GmbH
-All rights reserved. This program and the accompanying materials
-are made available under the terms of the Eclipse Public License v1.0
-which accompanies this distribution, and is available at
-http://www.eclipse.org/legal/epl-v10.html
-
-Contributors:
-   Ryan Henderson - initial API and implementation and/or initial
-   documentation
--->
-{% extends "layout.html" %}
-{% block body %}
-	<h4><b>Selected Visualization:</b> {{ current_vis }}</h4>
-  {% if settings is defined %}
-		{% for setting in settings %}
-			<p>
-        {{ setting }}: {{ settings[setting] }}
-			</p>
-		{% endfor %}
-  {% endif %}
-  {% block vis %}{% endblock vis %}
-  {% if duration is defined %}
-    <hr>
-    <p>Computing the visualization took {{ duration }} seconds.
-  {% endif %}
-  {% if reference_link is defined %}
-    <a href="{{ reference_link }}">Visualization Reference</a>
-  {% endif %}
-{% endblock %}
