--- picasso/templates/select_files.html
@@ -1,33 +1,0 @@
-<!--
-Copyright (c) 2017 Merantix GmbH
-All rights reserved. This program and the accompanying materials
-are made available under the terms of the Eclipse Public License v1.0
-which accompanies this distribution, and is available at
-http://www.eclipse.org/legal/epl-v10.html
-
-Contributors:
-   Ryan Henderson - initial API and implementation and/or initial
-   documentation
--->
-{% extends "layout.html" %}
-{% block body %}
-	<h4><b>Selected Visualization:</b> {{ current_vis }}</h4>
-  {% if settings is defined %}
-		{% for setting in settings %}
-			<p>
-        {{ setting }}: {{ settings[setting] }}
-			</p>
-		{% endfor %}
-  {% endif %}
-	<form action="{{ url_for('picasso.select_files') }}" method="POST" enctype="multipart/form-data">
-    <div class="form-group">
-      <label for="fileupload">Select input images</label>
-      <input type="file" name="file[]" multiple="" id="fileupload">
-      <p class="help-block">Image types only, please.  Will be resized to CNN input
-      shape.</p>
-      <label class="btn btn-primary">
-        Upload <input type="submit" value="Upload" style="display: none;">
-      </label>
-    </div>
-	</form>
-{% endblock %}
