--- picasso/templates/select_visualization.html
@@ -1,27 +1,0 @@
-<!--
-Copyright (c) 2017 Merantix GmbH
-All rights reserved. This program and the accompanying materials
-are made available under the terms of the Eclipse Public License v1.0
-which accompanies this distribution, and is available at
-http://www.eclipse.org/legal/epl-v10.html
-
-Contributors:
-   Ryan Henderson - initial API and implementation and/or initial
-   documentation
--->
-{% extends "layout.html" %}
-{% block body %}
-	<form name="choice" action="" method="POST">
-		<div class="form-group">
-			<label for="vis-select">Available visualizations</label>
-			<select class="form-control" name="choice" id="vis-select">
-				{% for vis in visualizations %}
-					<option value="{{ vis.name }}">{{ vis.description }}</option>
-				{% endfor %}
-			</select>
-		</div>
-		<div class="form-group">
-			<button class="btn btn-primary btn-block" type="submit">Go!</button>	
-		</div>
-	</form>
-{% endblock %}
