--- picasso/templates/settings.html
@@ -1,28 +1,0 @@
-<!--
-Copyright (c) 2017 Merantix GmbH
-All rights reserved. This program and the accompanying materials
-are made available under the terms of the Eclipse Public License v1.0
-which accompanies this distribution, and is available at
-http://www.eclipse.org/legal/epl-v10.html
-
-Contributors:
-   Ryan Henderson - initial API and implementation and/or initial
-   documentation
--->
-{% extends "layout.html" %}
-{% block body %}
-<h4><b>Selected Visualization:</b> {{ current_vis }}</h4>
-	<form name="settings" action="{{ url_for('picasso.select_files') }}" method="POST">
-		{% for setting in settings %}
-		<div class="form-group col-xs-2">
-			<label for="{{ setting }}">{{ setting }}</label>
-			<select class="form-control" name={{ setting }} id={{ setting }}>
-					{% for option in settings[setting] %}
-							<option value="{{ option }}">{{ option }}</option>
-					{% endfor %}
-				</select>
-		</div>
-		{% endfor %}
-			<button type="submit" class="btn btn-primary btn-lg btn-block">Choose files</button>
-	</form>
-{% endblock %}
